fix: change the location of the user information api call
