+++ src/pages/MyPage.tsx
@@ -4,20 +4,14 @@ import Profile from 'components/mypage/profile/Profile';
 import DefaultResume from 'components/mypage/defaultResume/DefaultResume';
 import { useMediaQuery } from 'react-responsive';
 import { useEffect } from 'react';
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
-import { SigninAtom, SigninStateAtom } from 'recoil/Signin';
-import { UserInfoAtom, UserProfileAtom } from 'recoil/UserProfile';
-import { InfoFormData } from 'data-type';
-import { GetCompanyProfile } from 'api/company_user';
-import { GetSeniorProfile } from 'api/senior_user';
+import { useRecoilValue } from 'recoil';
+import { SigninStateAtom } from 'recoil/Signin';
+import { UserInfoAtom } from 'recoil/UserProfile';
 import Suggestion from 'components/mypage/suggestion/Suggestion';
 import { useNavigate } from 'react-router-dom';
-import { GetProfile } from 'api/user';
 
 const MyPage = () => {
-  const signinData = useRecoilValue(SigninAtom);
-  const [userInfoData, setUserInfoData] = useRecoilState(UserInfoAtom);
-  const setUserProfileData = useSetRecoilState(UserProfileAtom);
+  const userInfoData = useRecoilValue(UserInfoAtom);
 
   const { isSignin } = useRecoilValue(SigninStateAtom);
   const navigate = useNavigate();
@@ -28,54 +22,6 @@ const MyPage = () => {
       navigate('/sign-in');
     }
   }, [isSignin]);
-
-  const getProfile = async (id: number) => {
-    const res = await GetProfile(id);
-    const data = res?.data;
-    setUserProfileData(data.image);
-  };
-
-  const getSeniorData = async (id: number) => {
-    const res = await GetSeniorProfile(id);
-    const data = res?.data;
-    saveData(data);
-  };
-
-  const getCompanyData = async (id: number) => {
-    const res = await GetCompanyProfile(id);
-    const data = res?.data;
-    saveData(data);
-  };
-
-  const saveData = (data: InfoFormData) => {
-    if (data) {
-      setUserInfoData({
-        id: data.id,
-        name: data.name,
-        username: data.username,
-        phone_number: data.phone_number,
-        email: data.email,
-        is_senior: data.is_senior,
-        is_enterprise: data.is_enterprise,
-        default_resume: data.default_resume ?? -1,
-        business_number: data.business_number ?? '',
-        is_certified: data.is_certified ?? false,
-      });
-    }
-  };
-
-  useEffect(() => {
-    const id = signinData.id;
-    const isSenior = signinData.is_senior;
-    if (id !== -1) {
-      getProfile(id);
-      if (isSenior) {
-        getSeniorData(id);
-      } else {
-        getCompanyData(id);
-      }
-    }
-  }, [signinData]);
 
   const isMobile = useMediaQuery({
     query: '(max-width:802px)',
